+++ pint_xarray/tests/test_conversion.py
@@ -356,6 +356,8 @@ class TestXarrayFunctions:
                 "x": ("x", x, {"units": original_units.get("x")}),
             },
         )
+        if type == "DataArray":
+            obj = obj["a"]
 
         if error is not None:
             with pytest.raises(error, match=match):
@@ -383,7 +385,6 @@ class TestXarrayFunctions:
         )
 
         if type == "DataArray":
-            obj = obj["a"]
             expected = expected["a"]
 
         actual = conversion.convert_units(obj, units)